+++ iot_firmware/enums.py
@@ -2,8 +2,13 @@ from enum import Enum
 from enum import EnumMeta
 
 
-class MetaEnum(EnumMeta):
-    def __contains__(cls, item):
+class NameClassMeta(type):
+    def __repr__(self) -> str:
+        return self.__name__
+
+
+class ContainsEnumMeta(EnumMeta):
+    def __contains__(cls, item) -> bool:
         try:
             cls(item)
         except ValueError:
@@ -11,9 +16,12 @@ class MetaEnum(EnumMeta):
         return True
 
 
-class StrEnum(str, Enum, metaclass=MetaEnum):
-    def __str__(self):
-        return "%s" % self.value
+class StrEnum(str, Enum, metaclass=ContainsEnumMeta):
+    def __str__(self) -> str:
+        return self.value
+
+    def __repr__(self) -> str:
+        return f"'{self.value}'"
 
 
 class MessageType(StrEnum):