+++ iot_firmware/event/enum.py
@@ -1,13 +1,23 @@
 """Event Enums."""
-import logging
-from enum import IntEnum
+from enum import auto
+from enum import Enum
+
+from iot_firmware.enums import StrEnum
 
 
-class EventLevel(IntEnum):
+class EventLevel(StrEnum):
     """Event levels."""
 
-    DEBUG = logging.DEBUG
-    INFO = logging.INFO
-    WARNING = logging.WARNING
-    ERROR = logging.ERROR
-    CRITICAL = logging.CRITICAL
+    DEBUG = "DEBUG"
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
+    CRITICAL = "CRITICAL"
+
+
+class EventHandlerState(Enum):
+    """EventHandler possible states."""
+
+    IDLE = auto()
+    RUNNING = auto()
+    STOPPING = auto()